pyproject.toml
[tool.poetry]
name = "devman-quiz-bot"
version = "0.1.0"
description = ""
authors = ["Stranix <Phantom2525@gmail.com>"]
readme = "README.md"
packages = [{include = "devman_quiz_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "13.12"
environs = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
